requirements.txt
pymongo>=4.5
python-dotenv>=1.0.0
rocketride
ipywidgets>=8.0.0
nest_asyncio>=1.5.0
jupyter>=1.0.0
